Moved user related apis in user blueprint from app.py
